+++ models/cdn.go
@@ -1,7 +1,7 @@
 package models
 
 type Image struct {
-	ID        uint   `json:"id" db:"id,pk,unique,autoinc"`
+	ID        int64  `json:"id" db:"id,pk,unique,autoinc"`
 	Source    string `json:"source" db:"source"`
 	Thumbnail string `json:"thumbnail" db:"thumbnail"`
 }
